Move wato plugin to datasource_programs file
