+++ cmk/gui/plugins/wato/datasource_programs.py
@@ -3599,3 +3599,28 @@ rulespec_registry.register(
         name="special_agents:rabbitmq",
         valuespec=_valuespec_special_agents_rabbitmq,
     ))
+
+
+def _valuespec_special_agents_fuse():
+    return Dictionary(
+        title=_("Fuse Management Central"),
+        help=
+        _("This rule set selects the special agent for Fuse Management Central "
+          "instead of the normal Check_MK agent and allows monitoring via REST API."),
+        elements=[
+            ("username", TextInput(title=_("Username"), allow_empty=False)),
+            ("password", Password(title=_("Password"), allow_empty=False)),
+            ("url", TextInput(title=_("Alerts API Endpoint"), allow_empty=False)),
+        ],
+        required_keys=["user","password","url"]
+    )
+
+
+rulespec_registry.register(
+    HostRulespec(
+        group=RulespecGroupDatasourceProgramsApps,
+        help_func=lambda: _("This rule selects the special agent for Fuse Management Central."),
+        name="special_agents:fuse",
+        title=lambda: _("Fuse Management Central"),
+        valuespec=_valuespec_special_agents_fuse,
+    ))
